Fix/queue ops (#25)
* Refactor KVStore initialization and add consumer setup in api-server

* Refactor API server to use MQConn and simplify consumer initialization

* Refactor transaction handling in API server to use MQConn for publishing transactions

* Update portfolio table schema: drop invested column and add avg_price column

* Refactor transaction model and update purse handling to use float for price and invested values

* Update NewRedis function to return error for connection failures

* Refactor transaction handling to push transactions to the queue in TradeService

* Refactor TradeService to integrate MQConn for transaction handling and update player pricing logic to use float values

* Refactor BallPicker to update player points using float values and improve error handling for Redis cache operations

* Refactor GetPlayerDetails to use float64 for BasePrice and CurPrice fields

* Add CacheService for managing player data and user balances with Redis caching

* Refactor Portfolio model and service to use float64 for price and balance fields

* Refactor leaderboard and league services to use float64 for price and points calculations

* Add health check endpoint and update CORS allowed origins; refactor database schema to use FLOAT for price fields

* Fix CORS allowed origins format and add blank line in schema.sql for better readability

+++ internals/leaderboard/leaderboard.go
@@ -49,7 +49,7 @@ func (l *Leaderboard) GetLeaderboard(leagueId string) ([]score, error) {
 		score.Points = resp.Balance
 		// Now go through each player from this user's portfolio and get their worth
 		for _, player := range resp.Players {
-			score.Points += player.CurPrice * player.Shares
+			score.Points += player.CurPrice * float64(player.Shares)
 		}
 		l.DB.Raw("SELECT user_name FROM users WHERE user_id = ?", userid).Scan(&score.UserName)
 		scores = append(scores, *score)
