+++ internals/leaderboard/models.go
@@ -1,7 +1,7 @@
 package leaderboard
 
 type score struct {
-	UserId   int    `json:"user_id"`
-	UserName string `json:"user_name"`
-	Points   int    `json:"points"`
+	UserId   int     `json:"user_id"`
+	UserName string  `json:"user_name"`
+	Points   float64 `json:"points"`
 }
